requirements.txt
coreapi
pytest
pytest-django
django
djangorestframework
django-environ